+++ lib/getch/gentoo/update.rb
@@ -4,11 +4,13 @@ require 'nito'
 
 module Getch
   module Gentoo
+    # update system gentoo
     class Update
       include NiTo
 
       def initialize
         @log = Log.new
+        @dest = "#{OPTIONS[:mountpoint]}/etc/portage"
         x
       end
 
@@ -18,12 +20,22 @@ module Getch
         sync
         add_musl_repo if OPTIONS[:musl]
         update
+        gpg
       end
 
       private
 
+      # https://wiki.gentoo.org/wiki/Gentoo_Binary_Host_Quickstart#Package_signature_verification
+      # Fix permissions error on gnupg directory
+      def gpg
+        return unless OPTIONS[:binary]
+
+        mv "#{@dest}/gnupg", "#{@dest}/gnupg.bak"
+        ChrootOutput.new('getuto')
+      end
+
       def sync
-        gentoo_conf = "#{OPTIONS[:mountpoint]}/etc/portage/repos.conf/gentoo.conf"
+        gentoo_conf = "#{@dest}/repos.conf/gentoo.conf"
         @log.info "Synchronize index, please waiting...\n"
         ChrootOutput.new('emaint sync --auto')
         sed gentoo_conf, /^sync-type/, 'sync-type = rsync'
@@ -31,8 +43,7 @@ module Getch
 
       def add_musl_repo
         Install.new('dev-vcs/git')
-
-        file = "#{OPTIONS[:mountpoint]}/etc/portage/repos.conf/musl.conf"
+        file = "#{@dest}/repos.conf/musl.conf"
         content = <<~CONF
         [musl]
         location = /var/db/repos/musl