+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 /boot partition, not /efi.
 * Add a new option `--root-size SIZE`.
 * Add a new option `--swap-size SIZE`.
+* Support [Gentoo binary](https://www.gentoo.org/news/2023/12/29/Gentoo-binary.html), use `--binary`, it make the install of Gentoo faster for small system.
 
 ## 0.7.3, release 2024-10
 Add a Workflow to build gem on Github